Add files via upload
The exam for the SN U1 will be done on computers and would be multiple choice questions answers (each question will have 4 options, one will be correct). There is a common database for these questions shared among Informatica, ACSAI and other courses. Students came together and formed a compilation of these questions with answers (some answers may be wrong). 

The original file was in ITA it is translated into EN using Doctranslate.com